🏷️ remove models.track, add models.constants
