+++ tiddl/models/resource.py
@@ -1,6 +1,57 @@
 from pydantic import BaseModel
 from datetime import datetime
 from typing import Optional, List, Literal, Dict
+from .constants import TrackQuality
+
+
+__all__ = ["Track", "Video", "Album", "Playlist"]
+
+
+class Track(BaseModel):
+
+    class Artist(BaseModel):
+        id: int
+        name: str
+        type: str
+        picture: Optional[str] = None
+
+    class Album(BaseModel):
+        id: int
+        title: str
+        cover: Optional[str] = None
+        vibrantColor: Optional[str] = None
+        videoCover: Optional[str] = None
+
+    id: int
+    title: str
+    duration: int
+    replayGain: float
+    peak: float
+    allowStreaming: bool
+    streamReady: bool
+    adSupportedStreamReady: bool
+    djReady: bool
+    stemReady: bool
+    streamStartDate: Optional[datetime] = None
+    premiumStreamingOnly: bool
+    trackNumber: int
+    volumeNumber: int
+    version: Optional[str] = None
+    popularity: int
+    copyright: str
+    bpm: Optional[int] = None
+    url: str
+    isrc: str
+    editable: bool
+    explicit: bool
+    audioQuality: TrackQuality
+    audioModes: List[str]
+    mediaMetadata: Dict[str, List[str]]
+    # for real, artist can be None?
+    artist: Optional[Artist] = None
+    artists: List[Artist]
+    album: Album
+    mixes: Dict[str, str]
 
 
 class Video(BaseModel):
@@ -53,7 +104,7 @@ class Album(BaseModel):
         picture: Optional[str] = None
 
     class MediaMetadata(BaseModel):
-        tags: List[Literal['LOSSLESS', 'HIRES_LOSSLESS']]
+        tags: List[Literal["LOSSLESS", "HIRES_LOSSLESS"]]
 
     id: int
     title: str
